refactor: split mdx server and client
